OT287-#042 order and rename index routes and components
